requirements.txt
numpy
Flask
imutils
opencv-python
Pillow
scikit-image
gunicorn